requirements.txt
requests
python-dotenv
docling
streamlit
langchain
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
transformers
langchain-groq